+++ src/image-3/utilities/get-type.ts
@@ -1,0 +1,10 @@
+export function getType(resource: any) {
+  if (resource['@type']) {
+    return resource['@type'];
+  }
+  if (resource.type) {
+    return resource.type;
+  }
+
+  return undefined;
+}