+++ src/index.umd.ts
@@ -1,7 +1,9 @@
 import * as Presentation2 from './presentation-2';
 import * as Presentation3 from './presentation-3';
+import * as Image3 from './image-3';
 
 export default {
   Presentation2,
   Presentation3,
+  Image3,
 };